+++ packages/integration-sdk-core/src/types/context.ts
@@ -2,8 +2,19 @@ import { IntegrationInstance, IntegrationInstanceConfig } from './instance';
 import { JobState } from './jobState';
 import { IntegrationLogger } from './logger';
 
+export type Execution = {
+  startedOn: number;
+  endedOn?: number;
+};
+
+export type ExecutionHistory = {
+  lastExecution: Execution;
+  lastSuccessfulExecution?: Execution;
+};
+
 export interface ExecutionContext {
   logger: IntegrationLogger;
+  history?: ExecutionHistory;
 }
 
 /**